CMake: align library-names between qmake and CMake
Install libqwebdav-qt6.so instead of a mixed-case name.
That way both build-methods yield the same .so name.

Note, though, that the specifics of the installed
namelinks are different:
- with CMake, the version is 1.0, and there are namelinks
  .so -> so.1 and .so.1 -> so.1.0 .
- with qmake, the version is 1.0.0, and there are links
  .so -> so.1.0.0 .so.1 -> so.1.0.0 and .so.1.0 -> so.1.0.0 .

Anything that is linked with unversioned, or .so version 1,
won't notice the difference. Only specific links to minor
or patch ABI versions are affected.